requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.1
idna==2.8
Markdown==3.0.1
oauthlib==3.0.1
Pillow==5.4.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
